+++ WinUtil/Encoding.hpp
@@ -17,10 +17,10 @@ namespace WinUtil
 		static std::string ToBase64(std::vector<unsigned char> const& bytes)
 		{
 			DWORD requiredChars = 0;
-			if (::CryptBinaryToStringA(&bytes[0], bytes.size(), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, NULL, &requiredChars))
+			if (::CryptBinaryToStringA(&bytes[0], static_cast<DWORD>(bytes.size()), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, NULL, &requiredChars))
 			{
 				std::vector<char> buffer(requiredChars, 0);
-				if (::CryptBinaryToStringA(&bytes[0], bytes.size(), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, &buffer[0], &requiredChars))
+				if (::CryptBinaryToStringA(&bytes[0], static_cast<DWORD>(bytes.size()), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, &buffer[0], &requiredChars))
 				{
 					return std::string(&buffer[0]);
 				}
@@ -37,7 +37,7 @@ namespace WinUtil
 		static std::vector<unsigned char> FromBase64(std::string const& base64)
 		{
 			std::vector<unsigned char> data(base64.size(), 0);
-			DWORD nBytes = data.size();
+			DWORD nBytes = static_cast<DWORD>(data.size());
 			if (::CryptStringToBinaryA(base64.c_str(), 0, CRYPT_STRING_BASE64, data.data(), &nBytes, NULL, NULL))
 			{
 				data.resize(nBytes);
